UPSTREAM: asus/am1i-a: Enable UART according to CONFIG_UART_FOR_CONSOLE
It has been observed by me and Elisenda Cuadros / Gergely Kiss [1] that
the boot process of this board is super slow when UART 0 is being used -
even if nothing is connected to it. Enable UART according to
CONFIG_UART_FOR_CONSOLE - and, if UART 0 is selected, it will be initialized
at romstage and this problem will not happen.

[1] https://mail.coreboot.org/pipermail/coreboot/2018-February/086132.html

BUG=none
BRANCH=none
TEST=none

Change-Id: I07c9cc94435045a6957f915a83052ea3f9b2f878
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 16a70c3d40db0d31d8b6c6c13603d27ad6bf5be3
Original-Signed-off-by: Mike Banon <mikebdp2@gmail.com>
Original-Change-Id: I6579aa8fd092da84f8afdcc33496db45c582919f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33796
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1683220
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>