UPSTREAM: libpayload/x86: Try to discover invariant TSC rate
We can skip the PIT-based TSC calibration if we can derive the invariant
TSC rate from CPUID/MSR data. This is necessary if the PIT is disabled,
which is the default, for instance, on Coffee Lake CPUs.

This implementation should cover all Intel Core i processors at least.
For older processors, we fall back to the PIT calibration.

BUG=none
BRANCH=none
TEST=none

Change-Id: I37b5704e1ee0c2b681bc2513dd737e759bbb0081
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bc2c12c72834a5b24fadd13d24df93298199dda1
Original-Change-Id: Ic6607ee2a8b41c2be9dc1bb4f1e23e652bb33889
Original-Signed-off-by: Nico Huber <nico.huber@secunet.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34170
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Michael Niewöhner <foss@mniewoehner.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2514341
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>