UPSTREAM: soc/intel/common/lpss: Add function to check for a LPSS controller
Add an API to check if device is a LPSS controller. This API can be
used for IRQ assignments for LPSS PCI controllers, since the LPSS
controllers have a requirement of unique IRQ assignments and do not
share same IRQ# with other LPSS controllers.

SOC code is reponsible to provide list of the LPSS controllers
supported and needs to implement soc_lpss_controllers_list API,
in case it needs to use this common implementation.

BUG=none
BRANCH=none
TEST=none

Change-Id: I534057260eb29391e2af217a396b96a78ada7ffa
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 141d90932396317074ea5c445d63e471f2907b7a
Original-Change-Id: I3f5bb268fc581280bb1b87b6b175a0299a24a44a
Original-Signed-off-by: Aamir Bohra <aamir.bohra@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34137
Original-Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1739989
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>