UPSTREAM: mb/siemens/{mc_apl1,...,mc_apl5}: Fix GPIO settings
Correct all GPIOs with reference to the Apollo Lake SoC EDS Vol 4
revision 2.4 chapter 10.1.2.3 List of Pins that are GPIOs but cannot be
used in Function 0 (GPIO) mode.

In additional, set an internal pull to any GPI that does not have an
external resistor so that the input is not in an undefined state.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia3ca1cdcfed79f7f7e46b54a851321f465f5b319
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2c7d1848856ce7bd8539ed4af460a476c39ff2fb
Original-Change-Id: Ia8fe457eddbed0f4ee6bff9ef9dd7a92545be40b
Original-Signed-off-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34379
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-by: Uwe Poeche <uwe.poeche@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1707304
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>