UPSTREAM: mb/google/eve: Enable wake from MKBP events in S3
We would like to wake eve up in suspend from an MKBP event.  This commit
simply enables MKBP events to wake the system in suspend using the
existing host event interface.  There is an accompanying series of
patches in the EC firmware for eve that will allow a MKBP wake mask to
be configured.

BUG=chromium:786721
BRANCH=firmware-eve-9584.B
TEST=Build and flash eve, generate MKBP events on the EC and verify
that the system wakes up in suspend.

Change-Id: I1f9eb601cf52ce9d3d2306ba88055c52bcabf05f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 19dca2b046d7acba126ee4403fd4c26df945d11c
Original-Change-Id: I75b05c83a4204d55df11589299a7488d04bbd073
Original-Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34454
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1712647
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>