UPSTREAM: soc/intel/{broad,cannon,sky}: Fix possible out-of-bounds reads
There will be a possible out of bounds array access if
power_limit_1_time == ARRAY_SIZE(power_limit_time_sec_to_msr), so
prevent that in the index check. This issue was fixed for other cpus in
commit 5cfef13f8d (cpu/intel: Fix out-of-bounds read due to off-by-one
in condition). Based on the discussion for that commit, also remove the
magic constant 28 in favour of the index of the last array element.

BUG=none
BRANCH=none
TEST=none

Change-Id: I830bdcca93f4d7e052b9ed48d7acc475885e9fb3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bcdb893778f857f310115522bbf7d70ad0cc017f
Original-Change-Id: Ic3f8735b23a368f8a9395757bd52c2c40088afa1
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Found-by: Coverity CID 1229673
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34498
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1726219
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>