UPSTREAM: src/device/oprom: Fix bootsplash display code for optionroms
So far the bootsplash is only correctly rendered if the framebuffer is
set up as 1024x768@16.
Different resolutions did not show anything, differnent depth resulted
in the distorted images.

This commit removes this limit by using the actual framebuffer resolutions
and combines the code for x86 and yabel.

For the moment the bootsplash is still limited to VGA-OptionROM
framebuffer init.

It was tested in 1280x1024@32 on the wip razer blade stealth using the
intel vgabios.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib24d3f2296084d92add9efb77d0ec66443f30282
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: db7a3ae8635a92764d357a93c04f49e9d9bbdca7
Original-Change-Id: I5ab7b8a0f28badaa16e25dbe807158870d06e26a
Original-Signed-off-by: Johanna Schander <coreboot@mimoja.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34537
Original-Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1721132
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>