UPSTREAM: vc/cavium/{bdk,include}: Clean up bdk_phys_to_ptr() calls
The bdk_phys_to_ptr() function converts a uint64_t address to a void *
pointer. Judging by the comments, the old implementation had a check
that would refuse to convert a null pointer, which required several
workarounds when trying to convert the address 0 to a pointer. This
isn't the case for coreboot though, which implements this function
as a simple (void *) cast, so we can remove the old workarounds.

BUG=none
BRANCH=none
TEST=none

Change-Id: I391d464796d9a481fa6b8e930cfa9f4a4cde872a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4926e989ac2f83bd887bee683c7e2c0481f5cd3a
Original-Change-Id: I6537d1699e6726c1fb155d69a51e14da856232de
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Found-by: Coverity CID 1393962
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34590
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1726218
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>