UPSTREAM: AUTHORS: Move src/acpi copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

BUG=none
BRANCH=none
TEST=none

Change-Id: Icc97bb0041680964c9a5a0f766781944c3d7c85c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 669e155ad2738c55e1bd52477a791afa682e23e9
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Id3382d19088cba2703350339b0bd0cfb3c0e63b2
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34604
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1726223
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>