UPSTREAM: AUTHORS: Move src/arch/arm64 copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

Additional changes in this patch:
- Make sure files say that they're part of the coreboot project
- Move descriptions below the license header

Note that the file include/arch/acpi.h is a fantastic example of why
moving to the authors file is needed.  Excluding the guard statements,
it has 8 lines of copyrights for 3 function declarations.

BUG=none
BRANCH=none
TEST=none

Change-Id: I3abd38411555a69b48611911a7e045423f2270ab
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 96c075a359a10cc9bae7a0630ae735868225a928
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I334baab2b4311eb1bd9ce3f67f49a68e8b73630c
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34606
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1771812
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>