UPSTREAM: Revert "src/security/vboot: Add option to skip display init with vboot 2.0"
This reverts commit 598af2e2c2785c00eb4290cdcefe1082b2a6f858.

Reason for revert: This commit breaks every board with VBOOT enabled
if the platform is apollolake, broadwell, skylake, baswell, baytrails
or icelake. The reason is, that the SoC selects
VBOOT_MUST_REQUEST_DISPLAY by default, and this has a dependency now
on VBOOT_MAY_SKIP_DISPLAY_INIT. This will only be auto-selected if
it is a CHROMEOS platform.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2986966874908c81b29f3a8a6cd5adc14fffa323
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 90cf4bb02aac15a41873ef9435d826b143a0a04e
Original-Change-Id: I3872d9aa993326ded135d8a5d950d5b1b1eddf34
Original-Signed-off-by: Christian Walter <christian.walter@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34308
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1726211
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>