UPSTREAM: soc/intel/common/gspi: Use GSPI bus id to map to the controller
Currently SPI bus id is used to map to the controller in order to set
the controller state. In certain platforms SPI bus id might not be
exactly the same as GSPI bus id. For example, in Intel platforms SPI bus
id 0 maps to fast spi i.e. SPI going to the flash and SPI bus id 1 .. n
map to GSPI bus id 0 .. n-1. Hence using SPI bus id leads to mapping to the
GSPI controller that is not enabled. Use the GSPI id bus so that the right
controller is set to active state. This fixes the regression introduced
by CB:34449

BUG=b:135941367
TEST=Boot to ChromeOS.

Change-Id: Iaeb76b2bc4aef2ba5f43edc9a6d4faa62e2a4966
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c6e37081749c3518a87a24b2b92bd9b0e293ebbf
Original-Change-Id: I792ab1fa6529f5317218896ad05321f8f17cedcd
Original-Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/34761
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Subrata Banik <subrata.banik@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1745342
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>